fix(trusty-git-analytics): extract jira client tests to fix 500-line cap
crates/trusty-git-analytics/src/collect/jira/client.rs was 667 SLOC,
over the 500-SLOC production cap, because its inline `#[cfg(test)] mod
tests` block (~270 lines covering changelog/comment parsing added for
#3966) doesn't match the test-file naming pattern the cap script uses
to grant the 1500-SLOC test budget. Extract the module to
client_tests.rs (loaded via `#[path = "client_tests.rs"] mod tests;`),
following the same split already used for
collect/github/client_tests.rs. No behavior change; all 9 extracted
tests still run and pass under the new module path.
